Rename style to be more specific
