+++ app/javascript/views/payments/paymentHistory/PaymentHistory.tsx
@@ -5,7 +5,7 @@ import { DownloadIcon } from "brave-ui/components/icons";
 import locale from "../../../locale/en";
 
 import { Header, Subheader } from "../../style";
-import { Cell, Link, Table, TableHeader } from "./style";
+import { Cell, Link, Table, TableHeader } from "./paymentHistoryStyle";
 
 export default class PaymentHistory extends React.Component {
   public render() {
