+++ app/javascript/views/payments/paymentsStyle.tsx
