+++ app/javascript/views/payments/upholdCard/UpholdCard.tsx
@@ -2,7 +2,8 @@ import * as React from "react";
 
 import { CaratRightIcon } from "brave-ui/components/icons";
 import { Header, Subheader } from "../../style";
-import { Text, Wrapper } from "./style";
+import { Text, Wrapper } from "./uploadCardStyle";
+
 import { UpholdStatus, WalletStatus } from "./upholdStatus/UpholdStatus";
 
 import locale from "../../../locale/en";
