+++ app/javascript/views/payments/upholdCard/upholdStatus/upholdStatusStyle.ts
