+++ app/javascript/views/payments/upholdCard/uploadCardStyle.ts
